requirements.txt
aiofiles==24.1.0
aiohttp==3.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
apng==0.3.4
astroid==2.5.1
asttokens==2.0.4
async-timeout==4.0.2
attrs==22.1.0
automationhat==0.2.0
beautifulsoup4==4.9.3
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==7.1.2
cloudscraper==1.2.71
colorama==0.4.4
colorzero==1.1
cryptography==3.3.2
dbus-python==1.2.16
discord.py==2.1.0
distro==1.9.0
distro-info==1.0+deb11u1
docutils==0.16
drumhat==0.1.0
envirophat==1.0.0
exceptiongroup==1.2.2
ExplorerHAT==0.4.2
Flask==1.1.2
fourletterphat==0.1.0
frozenlist==1.3.3
gpiozero==1.6.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==2.10
imageio==2.22.4
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jiter==0.9.0
lazy-object-proxy==0.0.0
logilab-common==1.8.1
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
multidict==6.0.3
mypy==0.812
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.2.2
openai==1.76.2
pantilthat==0.0.7
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera2==0.3.12
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==9.3.0
pixivpy3==3.7.5
psutil==5.8.0
pycairo==1.16.2
pycparser==2.21
pycups==2.0.1
pycurl==7.43.0.6
pydantic==2.11.4
pydantic-core==2.33.2
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
pyparsing==3.1.4
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5b0
pysmbc==1.0.23
python-apt==2.2.1.1
python-prctl==1.7
PyYAML==5.3.1
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.12.1
roman==2.0.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.6.0b1
sense-hat==2.6.0
simplejpeg==1.6.4
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
sniffio==1.3.1
soupsieve==2.2.1
spidev==3.5
ssh-import-id==5.10
thonny==4.0.1
toml==0.10.1
touchphat==0.0.1
tqdm==4.65.0
tweepy==4.12.1
twython==3.8.2
typed-ast==1.4.2
typing-extensions==4.12.2
typing-inspection==0.4.0
unattended-upgrades==0.1
unicornhathd==0.0.4
urllib3==2.2.2
v4l2-python3==0.3.2
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.8.2
